CXP-1439: 3 component split of Source Panel
