+++ components/catalogs/front/src/components/ProductMapping/components/SourceSelection/SourceSelection.tsx
@@ -1,0 +1,46 @@
+import React, {FC} from 'react';
+import {Source} from '../../models/Source';
+import {Target} from '../../models/Target';
+import {SourceErrors} from '../../models/SourceErrors';
+import {useAttribute} from '../../../../hooks/useAttribute';
+import {Attribute} from '../../../../models/Attribute';
+import {SectionTitle, Tag} from 'akeneo-design-system';
+import {useTranslate} from '@akeneo-pim-community/shared';
+import {SourceSettings} from './SourceSettings';
+import {createSourceFromAttribute} from '../../utils/createSourceFromAttribute';
+import {SelectSourceAttributeDropdown} from './SelectSourceAttributeDropdown';
+
+type Props = {
+    source: Source | null;
+    target: Target;
+    errors: SourceErrors | null;
+    onChange: (value: Source) => void;
+};
+
+export const SourceSelection: FC<Props> = ({source, target, errors, onChange}) => {
+    const translate = useTranslate();
+    const {data: attribute} = useAttribute(source?.source ?? '');
+
+    const handleSourceAttributeSelection = (selectedAttribute: Attribute) =>
+        onChange(createSourceFromAttribute(selectedAttribute));
+
+    return (
+        <>
+            <SectionTitle>
+                <Tag tint='purple'>1</Tag>
+                <SectionTitle.Title level='secondary'>
+                    {translate('akeneo_catalogs.product_mapping.source.title')}
+                </SectionTitle.Title>
+            </SectionTitle>
+            <SelectSourceAttributeDropdown
+                selectedCode={source?.source ?? ''}
+                target={target}
+                onChange={handleSourceAttributeSelection}
+                error={errors?.source}
+            />
+            {attribute !== undefined && source !== null && (
+                <SourceSettings source={source} attribute={attribute} errors={errors} onChange={onChange} />
+            )}
+        </>
+    );
+};
