+++ components/catalogs/front/src/components/ProductMapping/models/Source.d.ts
@@ -4,5 +4,5 @@ export type Source = {
     source: string | null;
     locale: string | null;
     scope: string | null;
-    parameters?: SourceParameter | null;
+    parameters?: SourceParameter;
 };
